Standardize unit interpretation of lookback_time in config for top_query_collection (#43618)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
As a standard, all type of interval values in collector config currently
requires an 's' suffix to interpret the unit. But the recently
introduced lookback_time in top_query_collection does not follow this
pattern. This needs to be fixed to maintain consistency and avoid
ambiguity in unit interpretation.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/43573

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Unit tests updated

+++ receiver/sqlserverreceiver/factory_test.go
@@ -43,7 +43,7 @@ func TestFactory(t *testing.T) {
 						InitialDelay:       time.Second,
 					},
 					TopQueryCollection: TopQueryCollection{
-						LookbackTime:        uint(2 * 10),
+						LookbackTime:        2 * 10,
 						MaxQuerySampleCount: 1000,
 						TopQueryCount:       200,
 						CollectionInterval:  time.Minute,
